Implemented credential reading and addressed PR comments
