+++ app/tooling/sales-admin/commands/vault.go
@@ -16,21 +16,27 @@ import (
 func Vault(vaultConfig vault.Config, keysFolder string) error {
 	vault, err := vault.New(vault.Config{
 		Address:   vaultConfig.Address,
-		Token:     vaultConfig.Token,
 		MountPath: vaultConfig.MountPath,
 	})
 	if err != nil {
 		return fmt.Errorf("constructing vault: %w", err)
 	}
 
-	if err := loadKeys(vault, vaultConfig.MountPath, os.DirFS(keysFolder)); err != nil {
+	initResponse, err := ReadCredentialsFile()
+	if err != nil {
+		return err
+	}
+
+	vault.SetToken(initResponse.RootToken)
+
+	if err := loadKeys(vault, os.DirFS(keysFolder)); err != nil {
 		return err
 	}
 
 	return nil
 }
 
-func loadKeys(vault *vault.Vault, mountPath string, fsys fs.FS) error {
+func loadKeys(vault *vault.Vault, fsys fs.FS) error {
 	fn := func(fileName string, dirEntry fs.DirEntry, err error) error {
 		if err != nil {
 			return fmt.Errorf("walkdir failure: %w", err)
